Add files via upload
Customer segmentation relies on identifying key differentiators that divide customers into groups that can be targeted. Clustering algorithms try to find natural clusters in data. Clustering can be done on the basis of Geography, spending, products, purchase history and then different deals/discounts can be offered to customers in different clusters.
The sole idea of clustering in this model is to identify target customers and develop different marketing strategies to increase sales and customer footfall.
The dataset is taken from Kaggle and it consisted of 200 rows and 5 features. The features includes: Customer ID, gender, age, spending score, age and annual income